Initial "Release"
This is just to have the files loaded and ready to modify more. These are BY NO MEANS ready to use. Feel free to help modify these to help get them working if you'd like. No need though, I plan to work on them more when I can, and or when I get more learned in the ways of coding! Thank you ♥